powertop: Add generic cpuidle C-state printing.
This adds C-state printing for non-intel platforms using a generic
cpuidle layout. This iterates through the number of online CPUs, descends
to their cpuidle/stateX/ directories and sets up a bitmap with all of the
possible C-states. On platforms that do not enable CX reporting via the
name file, this will just return.

Signed-off-by: Paul Mundt <lethal@linux-sh.org>